PR review - Reverted the custom stats interface, updated the generic stats format method, added required foreign field to stats object.
