+++ src/storage/json.ts
@@ -6,7 +6,6 @@ import type {
   TransactionStorage,
   SaveStats,
 } from "../types.js";
-import { saved } from "../messages.js";
 
 const logger = createLogger("LocalJsonStorage");
 
@@ -39,14 +38,12 @@ export class LocalJsonStorage implements TransactionStorage {
       total: txns.length,
       added: txns.length,
       updated: 0,
+      foreign: 0,
       pending: NaN,
       skipped: 0,
       existing: NaN,
     };
 
     return stats;
-  }
-  logStats(stats: SaveStats): string {
-    return saved(stats);
   }
 }
