+++ src/storage/sheets.ts
@@ -19,8 +19,7 @@ import type {
 } from "../types.js";
 import { TransactionStatuses } from "israeli-bank-scrapers/lib/transactions.js";
 import { sendDeprecationMessage } from "../notifier.js";
-import { normalizeCurrency } from "../utils/currency.js";
-import { saved } from "../messages.js";
+import { normalizeCurrency, checkIlsCharge } from "../utils/currency.js";
 
 const logger = createLogger("GoogleSheetsStorage");
 
@@ -108,6 +107,7 @@ export class GoogleSheetsStorage implements TransactionStorage {
       total: txns.length,
       added: 0,
       updated: 0,
+      foreign: 0,
       pending: 0,
       existing: 0,
       skipped: 0,
@@ -138,6 +138,12 @@ export class GoogleSheetsStorage implements TransactionStorage {
         }
 
         continue;
+      }
+
+      // Register foreign transactions that did not resolve to ILS yet (also pending)
+      const isIlsCharge = checkIlsCharge(tx);
+      if (!isIlsCharge) {
+        stats.foreign++;
       }
 
       if (tx.status === TransactionStatuses.Pending) {
@@ -203,8 +209,5 @@ export class GoogleSheetsStorage implements TransactionStorage {
     }
 
     this.sheet = doc.sheetsByTitle[worksheetName];
-  }
-  logStats(stats: SaveStats): string {
-    return saved(stats);
   }
 }
